Mars rover
Scientists are planning to send a "colleague" Tardy to the robot Wally. This is Billy's first landing on Mars, so he is being tested in a rectangular room measuring 15 × 20 m. The rover is landed in the center of the room (at point 8, 10), then control is transferred to the operator, that is, the user of your program.

The program asks which direction the operator wants to direct the robot: north (key W), south (key S), west (key A) or east (key D). The operator makes a choice, the rover moves in this direction by one meter, and the program reports the new position of the robot. If the rover hits a wall, it should not try to move towards the wall - in this case, its position does not change.