My checkpoint work on the SkillBox platform.
My checkpoint work on the SkillBox platform.

Task 1: Conversion
When paying for purchases with a card abroad, banks make conversion through an intermediate currency. For example, if you pay with a domestic card for goods in euros, then first this amount is converted into dollars, and then - into rubles.
What you need to do:
Write a program that receives as input the purchase price in euros and then outputs the answer in rubles. Let's imagine that we live in an alternate reality where 1 euro = 1.25 dollars and 1 dollar = 60.87 rubles.
............................................................
Task 2: Rough Math
In one center of mathematical analysis worked an intern who wrote programs for calculating functions. One day he was very tired and misunderstood the technical task, so the functions became roughly calculated.
His program works as follows: a sequence of N real numbers is entered, with positive numbers rounded up and negative numbers rounded down.

What you need to do:
Write a program that outputs the natural logarithm of a number if it is positive and the exponent in degree of the number if it is negative.
............................................................
Task 3. Steam analog
You are writing an installer program for a computer game. While the installer is downloading an update, you need to display the number of downloaded percentages for the user to understand whether he will have time to make tea before the process ends. Each game update requires a different number of megabytes, with different players having different internet connection speeds.
What you need to do:
Write a program that takes as input the size of the update file in megabytes and the speed of the internet connection in megabytes per second. For each second, the program should calculate and display the percentage of downloaded volume until the download is complete. At the end, the program should show how many seconds it took to download the update. Provide input control.
............................................................
Problem 4: The first digit
What needs to be done:
Given a positive number X, which consists of integer and fractional parts. Output its first digit after the decimal point. When solving this task, you cannot use conditional construction, loop or strings.
............................................................
Task 5: What volumes!
For a term paper in physics Andrei needs to compare the volumes of two planets: the Earth and a theoretically possible planet for our Universe. Andrei is good at formulas, but not good at math. He knows the volume of the Earth - it is 1.08321 * 10 ** 12 km³.

The volume of a theoretically possible planet he needs to calculate.
He has a formula: V = 43∙π∙R3 

In it, V is the volume, π is the number of pi, and R is the radius of the planet.

What you need to do:
Write a program that takes the radius of a random planet as input and displays how many times the planet Earth is smaller or larger than a theoretically possible planet in volume. Round the answer to three digits after the period.
............................................................