My checkpoint work on the SkillBox platform. 10.6
My checkpoint work on the SkillBox platform.

Task 1. Test task
Stepan gets a job and has to complete a test task: analyze a table, understand how it is built, and write a program to display it on the screen.

0	2	4	6	8	10	
1	3	5	7	9	11	
2	4	6	8	10	12	
3	5	7	9	11	13	
4	6	8	10	12	14	
5	7	9	11	13	15

What needs to be done:
Help Stepan implement the program.
Hint: pay attention to the column numbers and remember about the \t literal for tabulation.

Task 2. Stairs
What needs to be done:
Write a program that outputs a "stairs" of numbers when the user enters the number N.
To move numbers to a new line, it is allowed to use an empty print.

Task 3. Frame
What to do:
Write a program that draws a rectangular frame using symbolic graphics. For vertical lines, use the vertical bar (|), and for horizontal lines, use the hyphen (-) or underscore (_). Let the user define the width and height of the frame.

Task 4. Prime numbers
What you need to do:
Write a program that counts the number of prime numbers in a given sequence and displays the answer on the screen.
A prime number is divisible only by itself and by one. The sequence is specified by calling input on each iteration of the loop. One iteration is one number.

Task 5. The largest sum of digits
What to do:
The user enters N numbers. Among the natural numbers he has specified, find the largest sum of digits. Display this number and the sum of its digits.

Task 6. Triangle of hashtags
What you need to do:
Write a program that displays an isosceles triangle (pyramid) filled with hashtag symbols #. Let the user determine the height of the pyramid.